requirements.txt
python-escpos==2.2.0
tkcalendar==1.6.1
cryptography==3.4.8
